+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The error listener from antlr is removed since it caused problems, flooding stderr with messages - it will be later taken care of.
 - The goto definition provider logic was updated to provide proper goto definitions.
+- Server now requests the client to refresh semantic tokens on document saves and initialization.
 
 ### Changed